Appropriate results of gev(c|p)df function for small values of k
When computing the cdf and pdf of the GEV distributions, in functions
gevcdf and gevpdf, respectively, the parameter k appears in the
denominator of a division operation.  To avoid division by zero, the
condition k==0 is tested and an alternative formula is used for it.

However, this test should be extended to values very close to zero.
Doing some non-extensive tests, I concluded that the function returns
wrong results for values of abs(k) lesser than 1e-11.  The most
appropriate limit value should be more carefully chosen, though.

This problem was evidenced in unit tests of the gevcdf and gevpdf
functions.  In one case, we have:

    k = -0.03 : 0.01 : 0.03;

In my system (Octave 4.0 on Debian i386), the fourth element of k is
different from zero (it is equal 3.4694e-18) and the result of gevcdf
is clearly wrong, and the unit test fails.

hg:eff9575d05b15e436c4a59500f43f3fb6aa7f5a2