betastat: do not compare input against 'Inf' as computation will be NaN.
* betastat.m: there is no need to check if input is 'Inf' (with the aim
of finding the indicces to convert to NaN), since input values of Inf
will already compute a NaN anyway.  Also, reintroduce the use of find()
(which was dropped with 44787a23b69c).  Input 'a' and 'b' are likely to be
large, but the ammount of "invalid" (non-positive, Inf, or NaN) is likely
to be small.  In such cases, keeping the indices returned by find() is
actually more efficient.  Also add another test for handling of Inf and
NaN values.

hg:81c3668763e128039e72f96efe318020d67d9758