betastat: support for n dimensions and reduce memory usage for scalar input.
* betastat.m: dropping most of the input check automatically gives support
for arbitrary number of dimensions.  This also simplifies code.  In addition,
drop usage of common_size() which is not memory efficient.  If one of the
input is a scalar, we can use it as such, no need to create an array of the
repeated value.  Also add more tests for corner cases such as empty string,
empty array, and boolean values.

hg:ee92e5db9400b5f0c16e093d320e0fc9d90687a2