wishrnd fix
wishrnd takes a 'degrees of freedom' argument df.  When df
is an integer, wishrnd generates what is essentially a sum
of df semidefinite a rank one matrices.

When df is non-integral, the wishrnd uses a Bartlett
decomposition, generating several normal variates along with
chi2 variates with dof parameters

df, df - 1, ... , df - p + 1.

Where p is the dimension of the p x p scaling matrix Sigma.

Note that these are valid distributions as long as
df > p - 1.  The original code conditioned on df >= p, so I
am extending this condition to include fractional values
p - 1 < df < p.

The current behavior of the function is to truncate df to
floor(df) when the fractional df leads to an invalid
distribution.  This behavior is preserved, but I am adding a
warning so the user is informed that this is happening.